[fidl-cpp] explicitly delete move methods
This change explicitly deletes the move-assignment and move-constructor
of the fidl::Binding and fidl::BindingSet classes, and adds comments
indicating why. Prior to this change, use of std::move to create or
assign instances of these classes would result in a compilation error
referencing a match for the corresponding deleted non-move variants.
This can lead to a misunderstanding that the only reason these classes
do not support move semantics is that the methods are omitted from the
class definition.

Change-Id: I4698d14160e1b2f7d2968df63ec6b537f720c194