[DriverFramework][suspend] Fallback to reboot by default
Mostly Reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/307892
which was reverted due to unrelated failures https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=23108

The difference from previous changeset is that I am using inspector lib
instead of the task-utils. The unification of dump-thread from threads
uapp and libinspector needs a design and is out of scope for this
change.

When suspend task is stuck, currently we wait forever.
This is hindering infra workflows and so we should be rebooting
if the suspend timesout after 30sec. For debugging the driver suspend
issues, before kicking off the reboot, we dump the thread info of the
offending process.

ZX-4804:#Done Bug:34653 #Done

Test: Added automated test for suspend fail.
runtests -t devcoordinator-test. On Astro, I manually failed block
driver suspend, and ensured that the debug info is printed, with the
callstacks and dso list of the offending devhost.

Change-Id: Ibe01eb49630d5393a3301eb207454baa6fce657f