[skip-block] Attempt to read other partition copies.
The two astro partitions which depend on the copy functionality,
bootloader and sysconfig, both expect to use the entire partition. If a
bad block is grown, it becomes impossible to read. We should fall back
to other copies in that case, or in the case that a copy fails to be
read.

GetPartitionInfo() was also updated to reflect the potentionally larger
size.

Tested: runtests -t skip-block-test
Tested: Paved astro
Change-Id: If654158e98ec6f8bc5f6ac2d7c0843777520d816