[exception_broker][crashpad_agent] Use CrashReporter fidl api.
Before the communication between the Exception Broker and the Crashpad
Agent was through the fuchsia.crash.Analyzer api. Now we use the
fuchsia.feedback.CrashReporter.

This API expects a vmo with a minidump in it, so some functionality of
crashpad had to be inserted into the Exception Broker for this to work.
This functionality was removed from the Crashpad Agent.

SIZE CHANGES
From `fx blobstats` using for release build of astro.

       | Exception Broker | Crashpad Agent |  Total
-------|-------------------------------------------
Before |            98304 |         368640 | 466944
After  |           172032 |         278528 | 450650
-------|-------------------------------------------
Diff   |           +73728 |         -90112 | -16384

So this change saves 16kb of space.

TEST=Add unittest that calls the CrashReporter interface.
     Add unittest for creating VMO from the file reader.
     Add unittest to verify that the generated VMO is a valid one.
     generate the minidump data.
     `fx run-test exception_broker_tests`

     Run `fx shell -- crasher` with reporting uploading on.
     Generated report: 2713b3b33214cebce1

     Downloaded the minidump within that report and opened with zxdb.
     Opens file and am able to correctly see stack frames and source.

Change-Id: Idee6aafbd50540871763a4d6dce6793c6fa78cd7