[skip-block] Require only one partition copy write to succeed
The existing behavior would bail on attempting to write other copies if
a write to one of the copies fails. This is not ideal behavior for
astro, as we just need one of the copies to be written successfully to
update the bootloader or sysconfig partitions. This is particularly
important when we need to write the entire partition size, but are
unable to due to bad blocks.

Tested: runtests -t skip-block-test
Tested: Paved astro
Change-Id: I9f45af9a8641d6ccd7a3139a903e8c0e3496edd9