[audio_core] Adjust glitch logging display
This logging-only CL adjusts how we handle output underflows,
render underflows, and capture overflows, to articulate (display)
these in a consistent way.

Additionally, the threshold of what we consider a 'partial'
underflow has been adjusted to be four frames (previously was 1/2
frame). The snap-to-the-nearest-destination-frame step performed
at the beginning of each packet can (in normal non-underflow
cases) result in source-side shifts of as much as four frames, if
our sample-rate-conversion ratio is sufficiently large.

Finally, this fixes our terminology so that we refer to capture-
side data loss as overflow (not underflow).

Test: logging-only

Change-Id: I0bbbeb9b83e71a627e193f81a6a661cd73bcc0df