[fdio] _mmap_file: don't pass VMO_FLAG_EXEC
Right now, our expectation is that VMO_FLAG_EXEC in GetBuffer is not
load-bearing, which allows us to add policy restricting its use despite
callers potentially using mmap() with PROT_EXEC.

However, today this is load-bearing for Flutter/Dart.  To that end, we
work around this by replacing passing VMO_FLAG_EXEC to FileGetBuffer
with an explicit call to replace_as_executable for that VMO.  The
observed behavior is the same to the caller (because
replace_as_executable is still available ambiently, for now) but the
replace_as_executable call is made on the client side, rather than the
server side.

This allows us to make the server-side implementation enforce new
policy controls, without breaking clients.  When we've plumbed the
proper server-side implementation throughout, we can revert this change,
and everything should continue just working.  Without this change,
though, we can't write code that enforces policy without having first
shaved the entire yak of putting the ZX_FS_RIGHT_EXECUTABLE throughout
userspace, and we'd prefer to avoid allowing programs to take on new,
untracked dependencies on implicit executability in the meantime.

Bug: 35899
Change-Id: Ie9ca54eb4fc9466cd6b71a2f3b42a8c43613da81