[journal] Implement a simple version of revocation records
Instead of emitting a record on-disk, this implements a
simplified version of go/revoke, which updates the info block
when data writes overlap with metadata writes.

This allows for a "simpler" version of revocation records; as documented
in the codebase, a version which emits actual records may prove more
useful once metadata operations stop blocking the JournalWriter thread.

34525 #done

Change-Id: Icb2e0d731ca8aa2ba5d691efa91f982cc64b43cb