[shortcut] Improve shortcuts matching and registering
When registering a shortcut, SHIFT, ALT, CONTROL, and META could be used
to be notified of both LEFT_ and RIGHT_ variants of the key.

Also allows other modifiers to be enabled for triggering the shortcut,
for example enabled CAPS_LOCK will allow CTRL+C to be triggered.

Also, allow multiple shortcuts to be registered for the same listener.

Also, add a timeout when invoking a client listener.

Change-Id: I37f070c82f090698720761e22f1358e5cb0124cc