[stash] Add Flush() method to stash to wait on commit
The Commit() method on Stash queues an asynchronous atomic commit of the
stashed values, but it does not allow clients to observe when their
changes have been flushed to disk.

The new Flush() method has an empty result that can be waited on.
It simply commits and ensures that the stash has been written to disk.

In practice, we do not need to call sync_all on the underlying files
because the implementation seems to implicitly flush the file on close
before atomically copying it to the final location.

BUG: 35441
Change-Id: I14b58610c2c18e869e004b6ccd6fc1beb2040f82