[zircon][fvm][test]: Add devices library.
Move utest/fvm/utils.h to fvm/test/. Still seems a bit weird having
libraries within the test directory, but since they are test only is
not that bad.

This library provides wrappers for dealing with devices in tests,
detecting errors and such.

To preserve history, the classes are not split into their own headers,
this will happen in a follow up CL.

Unit tests for this testing utilities is blocked under a feature in
zxtest, for not bubbling every error up to the runner instance, such as
to allow testing test helpers like this one. This should come after a
possible refactor of zxtest to split into core/non-core tests.

NOTE: Consider renaming fs-utils to storage-test-utils.
Change-Id: Id87785c63668a6216ce743840bcc42061abd152e