[wlan] Add connection ping metric
Add connection ping metric and log two additional dimensions for
connection milestone metric.

The main difference between existing milestone metric and the new ping
metric is that the former logs the latest milestone when it happens,
while the latter periodically logs all milestones that have been reached
in the same session. The consequence is that in the daily per-device reports,
clients who connect longer than a day would show up in the first connected
milestone in the ping metric but not the milestone metric. This makes
the ping metric more useful for getting the daily count of connected devices.

Bug: 35400

Test:
1. fx run-test wlan-sme-tests
2. fx run-test wlanstack-tests
3. Connected to network and verified ping metric is logged periodically.
   Also verified that connection milestone metric is logged only when
   new milestone is reached.

Change-Id: I499954e4d43416f0ed93fde6b342367594e33563