[paver][bootserver] Add "--fail-fast-if-version-mismatch" to paver
The ota test added "--fail-fast" to make sure if the zedboot version
changes, we can exit early, rather than looping forever complaining about
the version mismatch. Unfortunately this also means that we disable the
paver handling packet loss, which makes this test more flaky. This adds
a new option "--fail-fast-if-version-mismatch" to the paver that only
exits early if the expected zedboot version doesn't match the device
zedboot, but otherwise allows retries.

Bug: 34872
Test: This is covered by the system ota tests
Change-Id: Ia8c571514ec00ae84f7fe27c691b3e4e0fd29db7