[systemupdater-tests] Verify paver service calls
This change switches the system updater integration tests to use the
real install-disk-image binary with a mocked paver service API that
verifies the expected interactions with the service.  A future change
will migrate the system updater to talk to the paver service directly,
and other than removing install-disk-image from this test package, these
tests should not need to be changed.

Bug: 35880
Change-Id: I3ffb6ff0254ede7addbcdb63c840d4e6be10dd6f