[perfcompare] Fix "Factor change" output field for zero or negative CI bounds
The "Factor change" field normally looks like the following, which
means that the factor change in the running time of this test was
somewhere between 0.395 and 2.484 (i.e. it was not clearly worse or
better):

  Factor change   Mean before              Mean after
  --------------  -----------------------  -----------------------
  0.395-2.484     60 +/- 5 ms              80 +/- 54 ms
  (from "fuchsia.ledger.get_page: cleared_pages/get_page")

However, before this CL, there are two problem cases:

1) If a confidence interval's lower bound goes negative, then "Factor
change" will contain a negative value such as -0.088 as follows (which
happens because the uncertainty of 42ms in the "Mean after" value is
greater than the estimated value 39ms):

  -0.088-3.977    27 +/- 6 ms              39 +/- 42 ms
  (from "fuchsia.ledger.get_page: new_page/get_page")

(That is an artifact of using t-test confidence intervals.  Those
allow the lower bound of the CI to come out negative, even though the
true values of most of our test statistics, like running times, can
never be negative.)

2) If the confidence intervals are both "0 +/- 0", then we get
"nan-nan" in the "Factor change" field:

  nan-nan         0 +/- 0 count            0 +/- 0 count
  (from "fuchsia.scenic.image_grid_cpp_stencil_shadow_volume: scenic_total_frames_dropped")

We fix those as follows:

 * Instead of "-0.088-3.977", output "ci_too_wide".
 * Instead of "nan-nan", output "no_change".

Split out FormatFactorRange() so that we can add a unit test for it.

This change also stops perfcompare.py from outputting the warning
"RuntimeWarning: invalid value encountered in double_scalars" (which
comes from doing division-by-zero on numpy.float64 values).

Bug: 35472
Test: "python perfcompare_test.py" (run by bots on CQ) + perfcompare trybot
Change-Id: If0cfc27251acc6949c46c80f203390d0943a38cd