[Clang] Temporarily disable -Wreorder-init-list warnings
https://reviews.llvm.org/D59754 limits basic designated initialiizers
for c++2a. The patch includes new warnings for when out of order,
nested, array, or mixed initialization is used.

Unfortunately, we still get the warnings even though we use
`-std=c++17`. While this is being addressed upstream, we should disable
the warning for now while fixing any instances of these invalid
initializations in the code base.

This helps unblock the next toolchain roll.

BUG: 35965
Change-Id: I4e1389085ad994185c320865b590b7ff9560a73e