[media][tests] Restructure fidelity testing recap
This test-only CL restructures how we present the fidelity results
summary. We no longer do this as a test case that just happens to be
last -- now this is done properly, after RUN_ALL_TESTS() returns.

This work is being done as part of larger changes (1) to relocate test
cases that are "unit-ish" into audio_mixer_unittests, and (2) to add
new fidelity-related tests in preparation of new resamplers landing.

Test: test-only

Change-Id: I45352418555bcf9f58c4fa4a3dfca37c163f050f