[kernel][object][port] Remove references to repeating waits in comments
There used to be a distinction between two different types of waits
set up using zx_object_wait_async():

 * ZX_WAIT_ASYNC_ONCE: one-shot/single-shot mode
 * ZX_WAIT_ASYNC_REPEATING: repeating mode

The latter was removed in cc6f71cadfc169d9745d424ffb75e27c3cea1498
(which removed the #define for the flag) and
4305907c5171525960bab8a8975eae8252dcf4a9 (which removed the rest of
the implementation).

We can remove the references to the distinction between the two modes.

Bug: none
Test: none
Change-Id: If2c178949ab2923a6c5ea59673b58a76ca392f44