[scenic] Flatland -- a 2D Layer API
This is a WIP CL for the 2D Layer API. It currently consists of a FIDL interface for a small number of functions, and a testable implementation.

This is intended to be a starting point for experimemtal iteration of all the various functions and dependencies needed to implement the full 2D Layer API. Until they are moved out of the fuchsia.ui.scenic.internal folder, the implementations should always be considered WIP (e.g., no on should make a dependnecy on this API). When they move out of that folder, they should go through both code and API review a second time.

TEST= fx shell runtests -t flatland_unittests

Change-Id: Iccf8b471dd6078775937c48ea25b2a1da1b8ff7c