[modular][future] Weak Maps flatten returned tuples
When a Future Map() returns a tuple, the arguments to the next function
in the chain are unbundled (flattened) from the tuple.

WeakMap was missing this capability. This CL applies the same solution
so WeakMaps now work the same. AsyncMap and WeakAsyncMap automatically
support their non-Async counterparts, so now all 4 variants work.

Unit tests are added that demonstrate the enhancement.

Test: fx run-test modular_lib_tests -t modular_lib_unittests -- --gtest_filter="FutureTest.*MapToTuple"
Change-Id: Ic9702e61c3f08ee31d9af30a86741e16d7d1559b