[ledger] Fix use-after-free in PageSync lifecycle
In case of an unrecoverable error, PageSync attempts to call both its idle
and error callbacks. If the former deletes the PageSync object, the latter
call will attempt to access freed member variables. This commit adds a
destruction sentinel to return early in that case and avoids triggering the
error callback.

Test: fx run-test ledger_tests -t ledger_unittests -- --gtest_filter=PageSyncImplTest.*
Change-Id: Ie93b5310e1b210f7835224af39a62d8952ff7dbe