[debugger] Support bitfields.
Adds support for reading and assignment of bitfields in collections
(structs and classes).

Bitfield information was removed from the BaseType. Even though DWARF
specifies this, types in the languages we support don't have bitfields,
only collection members.

Bitfield information is added to data members and the source information
for variables (so they can be written back later).

Change-Id: I92149ad66392b734de1cd45aca98f8e817f59c33