[debugger] Debug Agent won't attach twice to a process.
Before this it went and did all the attaching dance, only to fail down
the line when trying to listen to the exception port. It still returned
OK and led the client to believe it was attached twice. Now it returns
ZX_ERR_ALREADY_BOUND.

This creates some minor UI issues on the CLI that will be addressed in a
future CL.

TEST=Added unit tests.
     fx run-test debug_agent_tests

Change-Id: I354933a0df6f8ecdcba96ff868c37625d7683d45