[system-ota-test] connect to device after downloading artifacts
As the tests are defined, the OTA tests are executed right after the
device is paved. In 5b458bc85be6071d6dcffe4048e69c36aa6a2b53, I changed
the test behavior from connecting to the device as late as possible to be
as early as possible. Unfortunately this means the device might not have
finished booting before we try to connect, and we could time out. This
patch restores the old behavior of connecting as late as possible until
we implement a better solution to this condition.

Change-Id: I7730dbd76103041119d1cf0310a676fc885c229a