[sysconfig][sync-client] Create SyncClient to access syconfig partitions
Due to space reasons, it was necessary to pack five partitions into
a single nand block on astro. This library aims to make access to them
easier.

This library is anticipated to be used by two clients:
1) The paver for adding ABR functionality to astro.
2) The ChannelWriter fidl protocol implementation.

This functionality was not implemented as a new fidl service because it
was deemed unnecessary. Anything which has access to one partition
would likely have access to the other partitions even if each partition
was exposed as a seperate service. The additional complexity, however,
is high.

TESTED: runtests -t sysconfig-sync-client-test
Change-Id: I22a4b492a5f2be3e4cc76260ab9d6895b1e9ac4b