[netstack3] Add update_retain method to ConnAddrMap, IdMap
Add the `update_retain` method to the `ConnAddrMap` and `IdMap`
collection types. This method takes a callback which is invoked
with a mutable reference to each element in the collection in turn.
Those elements for which the callback returns `true` are kept, while
those elements for which the callback returns `false` are returned
via an iterator.

While we're here, also expand `ConnAddrMap` to be able to support
distinct address and connection objects. This prepares us for a
future change in which connections will have associated socket
objects that store more information beyond just the connection's
address.

Test: data_structures::id_map::tests::test_update_retain

Change-Id: Ibe1c317dee4a2c14d2bc6d9cbd3dab5628104e87