[sysmem][optee] sysmem support for VDEC secure mem
Some of this is based on jbauman@'s sysmem-assistant, which is now
removed in favor of this mechanism.

Exchange secure memory ranges between sysmem and TEE Controller,
including the VDEC range.

Create a paged heap for the VDEC range and serve that as a heap from
sysmem.

Test that a VDEC BufferCollection works, on astro and sherlock.

In the TEE Controller part of this, we use the TEEC_* client API in
loopback for consistency with other TEE client code.  Also, this should
make it easier to move this code in future if we decide we prefer to
serve fuchsia::sysmem::Tee from somewhere else.

Change-Id: Ie029bd0f8f0edb66d84ca82d657497c442a25c2e