[build] Allow Zircon's C/C++ libraries to be built in the GN build.
This change introduces an implementation of the "library" template that
works in the GN build.

This however does not modify any actual build yet. Instead, a parameter is
introduced to make "//zircon/public/lib/foobar" a simple alias for
"//zircon/system/ulib/foobar". Aliasing is disabled by default.

Things known to work with this change (and aliasing enabled):
- running a full build;
- deploying the result on hardware and not observing any boot issue;
- using SDK targets from //zircon/system/ulib (and not seeing any API change).

Things unknown after this change:
- impact on package/image sizes;
- differences in post-boot runtime behaviors (although CQ bots seemed happy).

Bug: 35659
Change-Id: I97dc4f41048dcc5b6e549dd07569f9feefc92153