[fuchsia-framebuffer] Support display buffering
Prior to this change, fuchsia-frambuffer was really only practical to use
with a single frame. Now there is a feature to be notified when a frame
is no longer being presented and is available to be updated and
presented again.

A new piece of sample code has been added to demonstrate the
new feature.

Testing: A new integration test creates a frame buffer and two frames and
makes sure that the new messages are properly delivered.

Change-Id: Idb9a5d4f1778026c2c501b378a5bad0d685245fd