[minfs] Reland "Enable journaling"
This is a reland of 564f531c542b2b5c2ed5d1572af55ce925872e55

Additionally updated non-FVM Minfs partitions to include an extra
block before the journal as a placeholder for the backup
superblock, and renamed this to the "integrity" section.
This greatly simplifies the transition to FVM / sparse files.

Original change's description:
> [minfs] Enable journaling
>
> Also allow upgrade from previous version of Minfs (7 or 8 -> 9).
>
> Change-Id: I009febe0e917f2081a6ec53a5217297f6ecf1fcf

Change-Id: I2ef7a7f0df95eb7b1e2c06b6dc9de1b0c85eaaea