[ledger] Tidy up PageManager::NewPageTracker
In the case of the was_opened_ vector being size one the was_opened_
vector is not guaranteed to contain the same operation_id as the
operation_id that is in scope at that point in the code, and
special-casing the vector-is-size-one case is no longer necessary
because following an earlier refactoring the vector no longer resides
in a map.

We also remove the special-casing of the vector-is-size-zero case
because it too became redundant with the .find call on the following
lines.

Testing removed lines of code is always hard; in the case of this
commit I'm not able at this time to design a regression test for this
content change. The vector-is-size-zero change would be completely
transparent to all tests. The removed statements in the
vector-is-size-one case were not what they were intended to be, and
thus were a bug, but it's not clear to me after a bit of tracing around
that they were a bug that would have been triggered consequent to
supported use of the PageManager class (the bug might have been
entirely latent). Now that this commit obviates the question of whether
the bug would ever have bitten, I'm reluctant spend much more thought
pursuing said question. :-)

Change-Id: I2552f6a2b52552a151ac08fdf0433c3b275bb16c