[debugger] Introduce ref counted thread suspension.
Before the thread would have only one suspend token, which would limit
the amount of interactions that breakpoints could do with threads.

This is made through tokens that have RAII-semantics related to the
debugged thread.

It was implemented as a parallel implementation because some users
expect different semantics and adding all in one change would make
reviewing harder.

TEST=Added a unit test.
     fx run-test debug_agent_tests

Change-Id: Ief7837607a6a5fa725c6f28721013c9004c17594