[kernel][vm] Log status in VmObjectPaged::GetPageLocked DEBUG_ASSERT
While trying to repro a different bug (36340) I saw an assert fire in
GetPageLocked, but it's unclear what went wrong because the status
code is not part of the assert message.

Extend the assert to emit the status code.

It's likely it was "just" an OOM.  Since callers of GetPageLocked must
already be prepared to handle ZX_ERR_NO_MEMORY there's really no
reason it should it crash the system.  Only assert if the error is not
ZX_ERR_NO_MEMORY.

Bug: 36340
Test: booted w/ 500mb, ran /boot/test/fs-test, saw no panics
Change-Id: I4d373de1e9a58b4098f92c2de129f93cee61b1c2