[debugger] Separate out DWARF DIE scanner.
This used to be part of the indexer which would iterate through all DIEs
in a unit and track parent information.

Separating this out allows the complex indexer code to be a little
simpler. More importantly it will allow this functionality to be re-used
more easily for a new indexer I'm planning to write.

TEST=Should be covered by existing indexer tests.

Change-Id: Idff239e912bced6915f2ff40076a35e97bc3d826