[fsl] Move garnet/public/lib/fsl to src/lib/fsl
This moves the contents of the directory:
  garnet/public/lib/fsl
to the directory:
  src/lib/fsl

to better match Fuchsia's desired source layout:
https://fuchsia.googlesource.com/fuchsia/+/refs/heads/master/docs/development/source_code/layout.md

This commit includes the following forwarding build targets to ease the
transition to the new layout:

  //garnet/public/lib/fsl:fsl
  //garnet/public/lib/fsl:fsl_tests

New code should rely on the new paths for these targets. A follow up commit
will remove the forwarding targets once all references are updated.

This commit includes the following forwarding headers to ease the
transition to the new layout:

  garnet/public/lib/fsl/handles/object_info.h
  garnet/public/lib/fsl/io/device_watcher.h
  garnet/public/lib/fsl/io/fd.h
  garnet/public/lib/fsl/io/redirection.h
  garnet/public/lib/fsl/socket/blocking_drain.h
  garnet/public/lib/fsl/socket/files.h
  garnet/public/lib/fsl/socket/socket_drainer.h
  garnet/public/lib/fsl/socket/strings.h
  garnet/public/lib/fsl/syslogger/init.h
  garnet/public/lib/fsl/tasks/fd_waiter.h
  garnet/public/lib/fsl/types/type_converters.h
  garnet/public/lib/fsl/vmo/file.h
  garnet/public/lib/fsl/vmo/shared_vmo.h
  garnet/public/lib/fsl/vmo/sized_vmo.h
  garnet/public/lib/fsl/vmo/strings.h
  garnet/public/lib/fsl/vmo/vector.h

New code should rely on the new paths for these headers. A follow up commit
will remove the forwarding headers once all references are updated.

This commit is generated by the script //scripts/move_source/move_source.py and
should contain no behavioral changes.

This also contains manually produced updates to documentation in:
  docs/development/languages/c-cpp/syslog.md
  src/developer/feedback/feedback_agent/tests/zx_object_util.h

Bug: 36063

Change-Id: I3ee88c5c507070f1858845c2c406c5fb400602d8