[scripts] Add a tool to help move source directories around
This adds a tool that helps move chunks of source code around. It
generates commits that move the source code while not modifying any
files outside the source and target directories. This allows landing
lots of source moves in parallel with each other and other changes.

A number of the functions in this script are likely to be useful in
source tree manipulation tools in the future.

Bug: 36063
Change-Id: Id57fe1e7cc960e3b803da3e0abc677b36190b6fd