[host-target-testing] bump connection timeout to 30 seconds
This gives a rebooted device more time to connect before timing out.
This can happen in CQ because devices are paved immedately before the OTA
tests are executed. If we try to connect to the device right when we start
then Fuchsia might not be fully booted by the time we try to connect.

Change-Id: Ib50f69ac011fd709d7884bfeda894e4ff2606191