[scenic] Add semaphores for double buffering swapchains
Modify FrameRecord so that each buffer can be associated with both a
pending frame and a finished frame.

Introduce a "retired" semaphore so that DrawAndPresentFrame can wait
for preceding frames to release the buffer. Allocate semaphores
exactly once, at startup.

Plumb a zx::event from DefaultFrameScheduler to DisplaySwapchain, so
that the frame scheduler can know when the swapchain has another
buffer free. For now, this does nothing. DefaultFrameScheduler can
perform admission control with this knowledge of frame queue depth.

SCN-1526 #comment

Test: fx aemu -N --software-gpu, also Astro and a Pixelbook
Test: added integration tests for display_swapchain.cc
Change-Id: Ib4e9bf0fbbf2ce8d8f4467c070f3c619435f1116