[fint] Initial integration interface commit
This change introduces the "fint" (*f*uchsia *int*egration) tool.
It is meant to drive fuchsia's continuous integration workflows
and realize the "integration interface".

The infrastructure will take an added dependency on a bootstapping
script, fint_init.sh, to build fint, so that it may in turn build the
platform.

Bug: 10428
Bug: 10429
Change-Id: Ied569140f0cac21d7a315e9eec3989bfd27816d0