[host-target-testing][system-ota-tests] connect to ssh in background
This patch fixes the condition where a device might not be fully booted
yet when we try to connect to it. Instead, `sshclient.Client` now makes
the initial connection in the background thread. In addition,
`sshclient.Client` was rewritten to support connection listeners, rather
than needing to poll on the `IsConnected()` method, and it grew a method
`WaitToBeConnected()` that blocks until the ssh client first connects.
Finally, `sshclient.Client` also was modified to block on connection, so
users of that library do not need to make an initial `WaitToBeConnected()`
call before using the library.

Change-Id: I053e2796b3d07a864a9f379d2de33ed686f1266f