[ledger] Obfuscate page ids in the cloud
Use a keyed hash to obfuscate the actual name of a page when syncing
through the cloud. The actual name is passed to Ledger through the
public API and can contain sensetive data.

This change also makes NewActivePageManager and CreatePageSync
asynchronous since we need to talk to the key service in order to
do keyed hashing.

BUG:35563

Change-Id: I8758eb0335535800d2adf3e3e2e32ee3daf539f6