[media.audio] Factor usage out of mixer/Gain
Because volumes must be combined prior to conversion to db, volumes must
become the unit we store for both of these.

This CL centralizes usage/stream gain interactions and updates in
StreamVolumeManager and adds tests for the logic.

This audit over updates to gain will make it possible to control
usage/stream in volume units instead of gain. It also removes checking
usage gain in the mix loop.

Bug: 36445
Bug: 36446
Bug: 35392
TEST: fx run-test audio_core_unittests
      fx run-test audio_mixer_test
      (Tests added for StreamVolumeManager)
Change-Id: I4ba38a6f45244c84be97803b6997c8db24441ea8