[pseudo-fs-mt] pseudo_directory! macro
It is often the case that we want to construct a small tree of
directories with a structure known at compile time.  While it is
possible to write code to do that using the io.fidl API, it requires a
lot of boilerplate code.

Similar to a `pseudo_directory!` macro in pseudo-fs library,
pseudo-fs-mt library users will benefit from a macro that would hide
most of the boilerplate.

ZX-4267 #done

Change-Id: Id7f85e6d6f53dc697d66f6f326add3c5a01153aa