[realm] use_parent_runners and have parent environment start runner
allow_parent_runners currently causes an environment to use its parent's
runner if the runner is already running. If it's not, the runner will be
started in the current environment, which doesn't allow for runners to
be shared between child environments.

This renames allow_parent_runners to use_parent_runners and guarantees
runners are set in an environment's parent if the option is set.

CF-758 #comment rename allow_parent_runners to use_parent_runners and
start runners in parent env

TEST=fx run-test appmgr_integration_tests

Change-Id: Id547b2d89ab1bc79216d0e0158399bb41583fcc9