[astro][securemem] Implement aml-securemem driver
This CL brings in a new driver to aid with grabbing physical addresses
of secure memory VMOs. In future CLs, this driver will also be able to
retrieve the region information of secure memory for sysmem.

Bug: 13689 #done

Test: New test has been added for bad cases. In order to test physical
VMO pinning, there will need to be additional work to introduce a
fake-vmo library.

Change-Id: I9683651597b1d9ae84b4b9f965c2d6502ff94aa4