[exception_broker] Introduce fuchsia.exception.ProcessLimbo.
ProcessLimbo is a service that is meant for components to query for any
processes that have been suspended on exception waiting for a handler to
appear.

This is the first version of the API that permits handles (such as a
debugger) to obtain exceptions that have occured on non-attached
processes.

This feature is not enabled by default and as of this CL it has to be
activated manually on source. A real activation mechanism will be
introduced on a later CL.

TEST=Added unit and integration (smoke) tests.
     fx run-tests exception_broker_tests
     Added the "limbo_caller" component that calls this services in
     order to verify that it's working. It's part of the
     debug_agent_tests package.
     $ run limbo_caller.cmx

Change-Id: Ie90f8648c298ef60bcdd3f6d564ca529676dd3e6