[kernel] Use generation counts in arena list
Changes the (head,next) pair of the HeadNode to be a (head,generation)
pair. Turns out (head, next) is still susceptible to ABA errors, it
just requires an extra node to be involved. A trace would like like:
 1. Start with a node chain of A->B->C->D->E. This means head_node_ is
    {AB}
 2. Enter GPArena::Alloc and proceed to just before the
    compare_exchange. Local expected head_node is {AB} and the planned
    next_head_node is {BC}
 3. Separate thread performs three allocations. Giving head_node_ at
    the end of {DE}
 4. Then frees nodes B and A giving chain A->B->D->E and head_node_ {AB}
 5. Thread 1 can now continue and do the compare and swap and now
    happily swaps {AB} for {BC}

Bug: 36268

Test: Writing a test that can execute threads with the precise
interleavings to trigger this is most likely impossible. Extensively
ran the reproduction scenario in 36268 and did not generate a failure
with this change.

Change-Id: Id299534b53bee7241963ba0db8968c96df8795b1