[refcount] Add implementation of BlockingRefCount.
A BlockingRefCount provides a thread-safe counter which can be
incremented and decremented, with an additional operation allowing
threads to wait for the count to become zero.

This can be useful in scenarios where an object is waiting for in-flight
callbacks to complete before cleaning up resources, for example:

  class MyClass {
    void PerformAsyncOperation() {
      // Increment the counter when we start the work, and decrement
      // it when finished.
      in_flight_ops_.Inc();
      DoWork(..., /*completion_callback=*/[this](){
        // in_flight_ops_.Dec();
      });
    }

    // Wait for all in-flight operations to terminate before
    // destructing.
    ~MyClass() {
      in_flight_ops_.WaitForZero();
    }

    BlockingRefCount in_flight_ops_;
  }

While this implementation isn't optimized for speed, it should avoid
syscalls in uncontended cases. If used more widely in
performance-sensitive areas, we can replace the implementation with
futexes without needing to change the interface.

Concretely, this will be used in the Intel HDA IPC code to avoid races
during object destruction.

Change-Id: I634039b3606cdb991103aa43b2048f7bf1e45fe2