[memory_metrics] Fixes problem with use of Cobalt API.
Also fixes logging.

The Cobalt API was not being used quite right and Cobalt was returning
an error but we were not seeing the error because logging wasn't working.

We also make a small enhancement to Cobalt's FakeLogger so that we
can use it to do a slightly deeper test in MetricsUnitTest that would
have caught this issue.

Change-Id: I011be02fcab3e62cb2c81c3e64dde8d2efa797c1