[kernel][x86] syscall: Remove unnecessary 'cli' in the sysret path
In the syscall return path, interrupts are disabled around the final
part of switching to a user stack frame, executing swapgs, and then
sysretq.

There is a direct path and an out-of-line path in x86_syscall()'s
return_from_syscall label; the direct path has interrupts disabled
on entry, the indirect path, handling thread signals, temporarily
enabled them while handling signals, and then jumps back to the
direct path.

Move the re-disablement of interrupts (CLI) to the out-of-line
path that re-enabled them. This makes the code clearer (STI/CLI
are balanced) and improves performance by removing an unneeded
CLI in the common case.

Tested:
* Booted on NUC7i3BNH (Kaby Lake)
* Booted on NUC7CJYH (Gemini Lake)
* Booted on AMD Ryzen R7 1700
* Booted on AMD 15h
* Booted on AMD 15h with the LK_DEBUGLEVEL-bracketed assert enabled

Change is performance-neutral on perfcompare (NUC7i3BNH)

Change is performance-positive on AMD Ryzen:
Summary counts:
716 test cases in total
711 test cases had no significant difference (no_sig_diff)
5 test cases got faster
0 test cases got slower
0 test cases added
0 test cases removed

https://docs.google.com/spreadsheets/d/10a57InA8bzWEH0ShKPPyWZ2kt24Pk34W9wsk0eVmVlw/edit#gid=0

Change-Id: I77c8ceb7a8d7c16735d73da24c8a2ae6dd16ed52