[intel-hda] Decouple DSP IPC notification logic from handling logic.
Currently, the IntelDspIpc class is responsible for both handling how
DSP notifications are received and what should be done with them.

This CL updates the class IntelDspIpc to just receive the notification
and trigger a callback, allowing external classses to actually perform
any logic required to deal with the notification. This helps to simplify
testing and, in future CLs, faking out the IPC code to help test
the rest of the DSP code.

Change-Id: Ie1ee5b1a4e2f889cd3610a592fc41e831df7d821