[x86] Allow C-states to be masked out at runtime
Introduces a per-CPU state mask which controls which C-states may be
entered for the processor in the idle loop. The mask can be controlled
at runtime via `k idlestates setmask`.

This mask defaults to only allowing C1/C1E, which is the current
behavior on all microarchitectures. A default can be provided on a
per-microarchitecture basis in x86/feature.cc.

This is useful for experimenting with different state selection policies
without needing to recompile across experiments.

Also reintroduce the state-set for Kaby Lake which was removed in
commit 993da30a64cb92e1c6f6932cb589f953462bf991.

Tested: k ut x86_idlestates; manual testing on a Kabylake device
Bug: 35289
Change-Id: I4693b2d98a9388598301a564bf0c5eb912b327e7