[system-ota-tests] refactor to make device initalization optional
Previously, the `pave => N-1 => N` and the `N => N'` OTAs were performed as
two separate subtests. This unfortunately results in a confusing error in
the test logs, where if the `pave => N-1 => N` subtest failed, the test
driver would still try to run the `N => N'` test. This CL changes the
system-ota-tests to only run the `pave => N-1 => N` if any of the
`--downgrade-*` arguments are specified. If not, it is assumed that the
device is already on `N` and it only runs the `N => N' => N` test.

Change-Id: I67e7e9fd8cd5dedfe12242278f055f676be4b039